refactor: move command abstract to abstractions dir
